Roll BoringSSL, update bindgen
The latest version of bindgen converts the C type `size_t` to a new
Rust type alias of the same name. On most platforms, this is an alias
to the unsigned integer type of platform's pointer width (e.g., `u32`
on 32-bit platforms, `u64` on 64-bit platforms, etc). Thus, we just
assume that that will be the case on all platforms we're compiled on,
and introduce some machinery to convert between `size_t` and `usize`
easily that will fail compilation if the conversion is not valid on
the target platform. See the `IntoSizeT` and `IntoUsize` traits in
the `boringssl::raw` module.

Change-Id: I2c22ba83de149587772668b8fd01a7f997604600
Reviewed-on: https://fuchsia-review.googlesource.com/c/mundane/+/375794
Reviewed-by: Drew Fisher <zarvox@google.com>