Pin to a particular version of bindgen
Changes to bindgen sometimes change the semantics of the generated
bindings (for example, by changing the Rust types that are generated
for particular C types). This change modifies `bindgen.sh` to "pin"
to a particular version of bindgen so that developers do not
accidentally use an older version of bindgen. Developers may still
use a more recent version of bindgen when one is available by editing
`bindgen.sh` to pin that more recent version.

Change-Id: Ibbd4878ca192e15ca0346cde23486d2d7311512f
Reviewed-on: https://fuchsia-review.googlesource.com/c/mundane/+/375795
Reviewed-by: Drew Fisher <zarvox@google.com>