Restrict use of BoringSSL's RSA setter API
The BoringSSL setter pattern, inherited from OpenSSL, is dangerous,
as it can cause keys to enter an invalid state without care around
how the keys are used. This change refactors the `rsa` module
slightly to make it less likely to accidentally make use of this
dangerous pattern. It also adds a comment explaining why this pattern
is dangerous.

Change-Id: Idb0eaf70460d7a26d81e933872e5050274af4550
Reviewed-on: https://fuchsia-review.googlesource.com/c/mundane/+/493422
Reviewed-by: Drew Fisher <zarvox@google.com>